refactor: address code review findings (A6, A7, P6, P7, T5)
- A6: Wire ContentRouter into cmd_login and output_response
- A7: Remove duplicate html_to_markdown/is_boilerplate from main.rs
- P6: Add 50MB PDF input size limit with clear error message
- P7: Detect encrypted PDFs and provide user-friendly error
- T5: Extract PdfChar/TextLine to shared content::types module
